fix workaround for javascript modules
