+++ package.json
@@ -11,6 +11,7 @@
   "keywords": [],
   "author": "Kent C. Dodds <kent@doddsfamily.us> (http://kentcdodds.com/)",
   "license": "MIT",
+  "esm": "auto",
   "dependencies": {
     "@polymer/lit-element": "^0.6.0",
     "@polymer/polymer": "^3.0.5",
